Update scsitape.1
An apostrophe at the beginning of the line is a command for (g)roff,
so it must be either escaped or moved away from the beginning of the
line to have a real apostrophe. The sequence \' doesn't stand for an
apostrophe, however, but an acute accent. There are escaped sequences
for an apostrophe, but they don't seem to be portable. Therefore, the
easiest solution is to move the apostrophe away from the beginning of
the line.

COC: Carsten Leonhardt <leo@debian.org>